Тестовый проект/requirements.txt
Flask==3.0.0
gunicorn==22.0.0

openai>=1.0.0,<2
python-dotenv>=1.0.0,<2
playwright>=1.41.0,<2


